C2C图片选择 照相功能的添加

+++ base/src/main/java/com/black/base/model/ProTokenResult.kt
@@ -2,5 +2,5 @@ package com.black.base.model
 
 class ProTokenResult {
     var proToken:String? = null
-    var expireTime:Long? = null
+    var expireTime:Int? = null
 }
